+++ app/models/claim.rb
@@ -91,6 +91,7 @@ class Claim < ApplicationRecord
   has_many :tasks
   has_many :amendments
   has_many :notes
+  has_one :support_ticket
 
   belongs_to :eligibility, polymorphic: true, inverse_of: :claim, dependent: :destroy
   accepts_nested_attributes_for :eligibility, update_only: true